Fixed bug where ThingsDatabase got purged when changing screen orientation
